alloc: the @string census closes, and a guard proves the count was never audited
r58a's counter shipped with one gap it named on itself: @string materialization
was deferred, because r57c was rewriting that file wholesale and taking it would
have meant a merge conflict across fourteen hunks. r57c has landed, so this is
the ordinary work DESIGN-allocation-counting.md §5 item 3 said should follow it
immediately.

Charged this commit, in string.cs unless stated: the materializing constructors
(ReadOnlySpan<byte>, IByteSeq<byte>, char[] -- which costs TWO, the intermediate
UTF-16 string and then the UTF-8 backing it encodes into), Encoding.UTF8's
GetBytes and GetString, ToRunes, the unsafe.StringData PinnedBuffer, the three
operator+ overloads, the copying []byte(s) conversions, the enumerator entry
points, and builtin.ToUTF8Bytes.

`this[Range]` -- Go's s[a:b] -- charges NOTHING, and that is a RESULT rather than
an omission: after r57c an @string is a window over shared backing, so slicing
one allocates nothing and copies nothing, exactly as in Go. Before that rewrite
this would have been the census's busiest site.

Three new AllocationCounter members carry the string shapes so counting stays a
SUBSTITUTION rather than an insertion, the same discipline NewArray/CopyOf
already enforce: NewString, Utf8ToString, Utf8ToBytes. Each charges nothing for
the empty case, because new string(char[0]), GetString(default) and GetBytes("")
all hand back interned or Array.Empty storage -- reporting an allocation that did
not happen is the same untruth as missing one that did.

THE GUARD FOUND A GAP THE CENSUS HAD MISSED, which is the reason it exists.
ByteSeqExtensions.ToSlice/ToGoString -- the []byte(s) and string(s) conversions
the converter emits for a `string | []byte`-constrained value -- materialized
their copies without charging them. A parseRFC3339-shaped body measured 192
B/parse and ZERO counted objects: arithmetic that cannot close, and precisely the
failure mode r56d warned about when it refused to land a partial counter. It is
charged now, and the shape reports exactly its six conversions.

GolibTests/AllocationCounterTests.cs is that guard. Every row of the census is
asserted as an EXACT object count rather than a range, and separately the
invariant the whole substitution rests on: a charged object costs at least the
CLR's 24-byte minimum, so count <= bytes/24 must hold across twelve shapes, or
AllocsPerRun could report a number that rises rather than falls and turn a
passing assert into a failing one. The doc comment on NewArray had claimed such a
check existed since r58a; it did not, and that is exactly how a census drifts.

ByteSeqAllocationTests' @string bound was the handoff r57c left for this lane. It
asserted < 600 B/parse against a figure the window fix had already moved to 192,
so it guarded nothing. It now asserts the COUNT -- exactly six objects, one per
[]byte(s) conversion parseUint performs -- because the count is the quantity that
is actually invariant here: the byte figure moves with the input's length, the
number of copies does not. The byte bound stays as its cross-check, retightened
to 250. The history of that one number is the history of both fixes: 776 B/parse
originally, 416 after the IByteSeq redesign removed the boxing, 192 once
sub-slicing stopped copying.

The stale "@string materialization is outside the census" clause is removed from
AllocationCounter's coverage statement, AllocsPerRun's documentation, and both
runtime notes, since it is no longer true.

Also folded in, from the same lane's unfinished work: array<T>(ulong) was routed
through NewArray<T>(nint), and the narrowing cast was a behavior change nobody
asked for. `new T[someHugeUlong]` throws OutOfMemoryException; the cast wraps it
negative first, so it threw something else instead. A ulong overload takes the
length unchanged and restores exactly the exception that site raised before the
counter existed -- the counter is supposed to observe allocation, not alter it.

+++ src/core/golib/array.cs
@@ -75,7 +75,7 @@ public readonly struct array<T> : IArray<T>, IList<T>, IReadOnlyList<T>, IEquata
 
     public array(ulong length)
     {
-        m_array = AllocationCounter.NewArray<T>((nint)length);
+        m_array = AllocationCounter.NewArray<T>(length);
         m_length = (int)length;
     }
